Remove vote and result saving from functions
test_save_voting_round is split up into test_save_scores test_save_votes.

The dependency on functions and voting_round is removed from score saving tests.
The dependency on functions, score and score_helpers is removed from vote saving tests.
