+++ src/test/CMakeLists.txt
@@ -139,15 +139,23 @@ add_executable(${test_name}
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
 
-set(test_name test_save_voting_round)
+set(test_name test_save_scores)
 add_executable(${test_name}
 	${test_name}.cpp
 	testing.cpp
-	${PROJECT_SOURCE_DIR}/src/functions.cpp			# TODO: Remove dependency to this file here
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
-	${PROJECT_SOURCE_DIR}/src/score.cpp				# TODO: Remove dependency to this file here
-	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp		# TODO: Remove dependency to this file here
+	${PROJECT_SOURCE_DIR}/src/score.cpp
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp)
+target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
+add_test(NAME ${test_name} COMMAND ${test_name})
+
+set(test_name test_save_votes)
+add_executable(${test_name}
+	${test_name}.cpp
+	testing.cpp
+	${PROJECT_SOURCE_DIR}/src/helpers.cpp
+	${PROJECT_SOURCE_DIR}/src/print.cpp
 	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
