+++ src/test/test_save_scores.cpp
@@ -1,0 +1,39 @@
+#include <filesystem>
+#include <fstream>
+
+#include "functions.h"
+#include "score_helpers.h"
+#include "testing.h"
+
+namespace
+{
+
+constexpr auto kTestFileName = "placeholder_file_name.txt";
+
+void saveWhenNoScores() {
+	std::filesystem::remove(kTestFileName);
+
+	ASSERT_FALSE(saveScores({}, kTestFileName));
+	ASSERT_FALSE(std::filesystem::exists(kTestFileName));
+}
+void saveWhenSomeScores() {
+	std::filesystem::remove(kTestFileName);
+
+	Scores const scores{ Score{ "item1", 3, 4 }, Score{ "item3", 1, 3 } , Score{ "item2", 7, 0 } };
+	ASSERT_TRUE(saveScores(scores, kTestFileName));
+	ASSERT_TRUE(std::filesystem::exists(kTestFileName));
+
+	std::filesystem::remove(kTestFileName);
+}
+
+} // namespace
+
+int main() {
+	std::filesystem::remove(kTestFileName);
+
+	saveWhenNoScores();
+	saveWhenSomeScores();
+
+	std::filesystem::remove(kTestFileName);
+	return 0;
+}
